Refactor input system: remove duplicates
