+++ src/input/CMakeLists.txt
@@ -1,9 +1,8 @@
 add_library(input STATIC
-    input.hpp
-    input.cpp
     input_manager.hpp
     input_manager.cpp
     key_constants.hpp
+    key_constants.cpp
 )
 
 target_include_directories(input PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
